+++ dagger2/src/main/kotlin/binds/AppleDatabase.kt
@@ -1,4 +1,4 @@
-package database
+package binds
 
 import javax.inject.Inject
