+++ dagger2/src/main/kotlin/binds/DatabaseComponent.kt
@@ -1,4 +1,4 @@
-package database
+package binds
 
 import dagger.Component
