+++ dagger2/src/main/kotlin/binds/DatabaseInterface.kt
@@ -1,4 +1,4 @@
-package database
+package binds
 
 interface DatabaseInterface {
     fun loadMessage(): String
