+++ dagger2/src/main/kotlin/constructor/Calculator.kt
@@ -1,4 +1,4 @@
-package calculator
+package constructor
 
 import javax.inject.Inject
