+++ dagger2/src/main/kotlin/constructor/CalculatorComponent.kt
@@ -1,4 +1,4 @@
-package calculator
+package constructor
 
 import dagger.Component
