+++ dagger2/src/main/kotlin/fieldinjection/Apple.kt
@@ -1,0 +1,7 @@
+package fieldinjection
+
+import javax.inject.Inject
+
+class Apple @Inject constructor(){
+    val name: String = "Apple"
+}
