+++ dagger2/src/main/kotlin/fieldinjection/FruitsApplication.kt
@@ -1,0 +1,15 @@
+package fieldinjection
+
+import javax.inject.Inject
+
+class FruitsApplication {
+    @Inject
+    lateinit var apple: Apple
+
+    init {
+        DaggerFruitsComponent
+            .builder()
+            .build()
+            .inject(this)
+    }
+}
