+++ dagger2/src/main/kotlin/fieldinjection/FruitsComponent.kt
@@ -1,0 +1,12 @@
+package fieldinjection
+
+import dagger.Component
+
+
+/**
+ * フィールドインジェクション
+ */
+@Component
+interface FruitsComponent {
+    fun inject(application: FruitsApplication)
+}
