+++ dagger2/src/main/kotlin/provides/CarComponent.kt
@@ -1,4 +1,4 @@
-package car
+package provides
 
 import dagger.Component
