+++ dagger2/src/main/kotlin/provides/Engine.kt
@@ -1,4 +1,4 @@
-package car
+package provides
 
 interface Engine {
     val speed: Int
