+++ dagger2/src/main/kotlin/qualifier/PhoneComponent.kt
@@ -1,4 +1,4 @@
-package phone
+package qualifier
 
 import dagger.Component
