+++ dagger2/src/main/kotlin/qualifier/PhoneModule.kt
@@ -1,4 +1,4 @@
-package phone
+package qualifier
 
 import dagger.Module
 import dagger.Provides
